split up draft methods
