+++ src/main/java/ti4/commands/milty/StartMilty.java
@@ -13,6 +13,7 @@ import ti4.map.Game;
 import ti4.message.MessageHelper;
 import ti4.model.MapTemplateModel;
 import ti4.model.Source.ComponentSource;
+import ti4.service.draft.DraftSpec;
 import ti4.service.milty.MiltyService;
 
 class StartMilty extends GameStateSubcommand {
@@ -31,7 +32,7 @@ class StartMilty extends GameStateSubcommand {
     @Override
     public void execute(SlashCommandInteractionEvent event) {
         Game game = getGame();
-        MiltyService.DraftSpec specs = new MiltyService.DraftSpec(game);
+        DraftSpec specs = new DraftSpec(game);
 
         // Map Template ---------------------------------------------------------------------------
         MapTemplateModel template = getMapTemplateFromOption(event, game);
