+++ src/main/java/ti4/draft/StandardBagDraft.java
@@ -13,8 +13,7 @@ import ti4.map.Game;
 import ti4.message.logging.BotLogger;
 import ti4.message.logging.LogOrigin;
 import ti4.model.FactionModel;
-import ti4.service.milty.MiltyDraftHelper;
-import ti4.service.milty.MiltyDraftManager;
+import ti4.service.draft.DraftTileManager;
 
 public class StandardBagDraft extends BagDraft {
     public StandardBagDraft(Game owner) {
@@ -73,10 +72,10 @@ public class StandardBagDraft extends BagDraft {
                 DraftItem.Category.HOMESYSTEM, HomeSystemDraftItem.buildAllDraftableItems(allDraftableFactions));
         allDraftableItems.put(DraftItem.Category.DRAFTORDER, SpeakerOrderDraftItem.buildAllDraftableItems(game));
 
-        MiltyDraftManager draftManager = game.getMiltyDraftManager();
-        MiltyDraftHelper.initDraftTiles(draftManager, game);
-        allDraftableItems.put(DraftItem.Category.REDTILE, RedTileDraftItem.buildAllDraftableItems(draftManager));
-        allDraftableItems.put(DraftItem.Category.BLUETILE, BlueTileDraftItem.buildAllDraftableItems(draftManager));
+        DraftTileManager.addAllDraftTiles(game);
+        DraftTileManager draftTileManager = game.getDraftTileManager();
+        allDraftableItems.put(DraftItem.Category.REDTILE, RedTileDraftItem.buildAllDraftableItems(draftTileManager));
+        allDraftableItems.put(DraftItem.Category.BLUETILE, BlueTileDraftItem.buildAllDraftableItems(draftTileManager));
 
         List<DraftBag> bags = new ArrayList<>();
