+++ src/main/java/ti4/draft/items/RedTileDraftItem.java
@@ -14,9 +14,9 @@ import ti4.model.PlanetModel;
 import ti4.model.PlanetTypeModel;
 import ti4.model.TechSpecialtyModel;
 import ti4.model.TileModel;
+import ti4.service.draft.DraftTileManager;
 import ti4.service.emoji.MiscEmojis;
 import ti4.service.emoji.TI4Emoji;
-import ti4.service.milty.MiltyDraftManager;
 import ti4.service.milty.MiltyDraftTile;
 
 public class RedTileDraftItem extends DraftItem {
@@ -80,19 +80,19 @@ public class RedTileDraftItem extends DraftItem {
         return MiscEmojis.Supernova;
     }
 
-    public static List<DraftItem> buildAllDraftableItems(MiltyDraftManager draftManager) {
+    public static List<DraftItem> buildAllDraftableItems(DraftTileManager draftTileManager) {
         List<DraftItem> allItems = new ArrayList<>();
-        for (MiltyDraftTile tile : draftManager.getRed()) {
+        for (MiltyDraftTile tile : draftTileManager.getRed()) {
             allItems.add(generate(Category.REDTILE, tile.getTile().getTileID()));
         }
         DraftErrataModel.filterUndraftablesAndShuffle(allItems, Category.REDTILE);
         return allItems;
     }
 
-    public static List<DraftItem> buildAllDraftableItems(MiltyDraftManager draftManager, Game game) {
+    public static List<DraftItem> buildAllDraftableItems(DraftTileManager draftTileManager, Game game) {
         List<DraftItem> allItems = new ArrayList<>();
         String[] results = PatternHelper.FIN_SEPERATOR_PATTERN.split(game.getStoredValue("bannedTiles"));
-        for (MiltyDraftTile tile : draftManager.getRed()) {
+        for (MiltyDraftTile tile : draftTileManager.getRed()) {
             if (Arrays.asList(results).contains(tile.getTile().getTileID())) {
                 continue;
             }
