+++ src/main/java/ti4/service/milty/MiltyDraftHelper.java
@@ -8,10 +8,7 @@ import java.awt.Point;
 import java.awt.image.BufferedImage;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.entities.channel.middleman.MessageChannel;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
@@ -23,20 +20,16 @@ import ti4.image.DrawingUtil;
 import ti4.image.ImageHelper;
 import ti4.image.MapGenerator.HorizontalAlign;
 import ti4.image.Mapper;
-import ti4.image.TileHelper;
 import ti4.map.Game;
 import ti4.map.Planet;
 import ti4.map.Player;
-import ti4.map.Tile;
 import ti4.map.UnitHolder;
 import ti4.message.MessageHelper;
 import ti4.message.logging.BotLogger;
 import ti4.model.FactionModel;
 import ti4.model.MapTemplateModel;
 import ti4.model.Source.ComponentSource;
-import ti4.model.TileModel;
-import ti4.model.TileModel.TileBack;
-import ti4.model.WormholeModel;
+import ti4.service.draft.DraftSliceHelper;
 import ti4.service.emoji.MiscEmojis;
 import ti4.service.emoji.TI4Emoji;
 import ti4.service.emoji.TechEmojis;
@@ -45,8 +38,6 @@ import ti4.service.milty.MiltyDraftManager.PlayerDraft;
 
 @UtilityClass
 public class MiltyDraftHelper {
-
-    private static final Map<String, MiltyDraftTile> tiles = new HashMap<>();
 
     public static void generateAndPostSlices(Game game) {
         MessageChannel mainGameChannel = game.getMainGameChannel();
@@ -288,88 +279,6 @@ public class MiltyDraftHelper {
         return ImageHelper.readEmojiImageScaled(emojiString, 40);
     }
 
-    public static void initDraftTiles(MiltyDraftManager manager, Game game) {
-        List<ComponentSource> sources = new ArrayList<>(Arrays.asList(
-                ComponentSource.base,
-                ComponentSource.codex1,
-                ComponentSource.codex2,
-                ComponentSource.codex3,
-                ComponentSource.codex4,
-                ComponentSource.pok));
-        if (game.isDiscordantStarsMode() || game.isUnchartedSpaceStuff()) {
-            sources.add(ComponentSource.ds);
-            sources.add(ComponentSource.uncharted_space);
-        }
-        initDraftTiles(manager, sources);
-    }
-
-    private static MiltyDraftTile getDraftTileFromModel(TileModel tileModel) {
-        String tileID = tileModel.getId();
-        if (tiles.containsKey(tileID)) return tiles.get(tileID);
-
-        Set<WormholeModel.Wormhole> wormholes = tileModel.getWormholes();
-        MiltyDraftTile draftTile = new MiltyDraftTile();
-        if (wormholes != null) {
-            for (WormholeModel.Wormhole wormhole : wormholes) {
-                if (wormhole == WormholeModel.Wormhole.ALPHA) {
-                    draftTile.setHasAlphaWH(true);
-                } else if (wormhole == WormholeModel.Wormhole.BETA) {
-                    draftTile.setHasBetaWH(true);
-                } else {
-                    draftTile.setHasOtherWH(true);
-                }
-            }
-        }
-
-        Tile tile = new Tile(tileID, "none");
-        draftTile.setTile(tile);
-
-        for (Planet planet : tile.getPlanetUnitHolders()) {
-            draftTile.addPlanet(planet);
-        }
-
-        if (tile.isAnomaly()) {
-            draftTile.setTierList(TierList.anomaly);
-        } else if (tile.getPlanetUnitHolders().isEmpty()) {
-            draftTile.setTierList(TierList.red);
-        } else {
-            draftTile.setTierList(TierList.high);
-        }
-
-        tiles.put(tileID, draftTile);
-        return draftTile;
-    }
-
-    public static void initDraftTiles(MiltyDraftManager draftManager, List<ComponentSource> sources) {
-        List<TileModel> allTiles = new ArrayList<>(TileHelper.getAllTileModels());
-        for (TileModel tileModel : allTiles) {
-
-            if (isInvalid(tileModel)) continue;
-            //  System.out.println(tileModel.getSource() + tileModel.getName());
-            if (!sources.contains(tileModel.getSource())) continue;
-            if (tileModel.getTileBack() == TileBack.GREEN || tileModel.isHyperlane()) continue;
-
-            MiltyDraftTile draftTile = getDraftTileFromModel(tileModel);
-            draftManager.addDraftTile(draftTile);
-            // System.out.println("yes");
-        }
-    }
-
-    private static boolean isInvalid(TileModel tileModel) {
-        TileModel.TileBack back = tileModel.getTileBack();
-        if (back != TileBack.RED && back != TileBack.BLUE) {
-            return true;
-        }
-
-        String id = tileModel.getId().toLowerCase();
-        String path =
-                tileModel.getImagePath() == null ? "" : tileModel.getImagePath().toLowerCase();
-        List<String> disallowedTerms = List.of(
-                "corner", "lane", "mecatol", "blank", "border", "fow", "anomaly", "deltawh", "seed", "mr", "mallice",
-                "ethan", "prison", "kwon", "home", "hs", "red", "blue", "green", "gray", "gate", "setup");
-        return disallowedTerms.stream().anyMatch(term -> id.contains(term) || path.contains(term));
-    }
-
     public static void buildPartialMap(Game game, GenericInteractionCreateEvent event) throws Exception {
         MiltyDraftManager manager = game.getMiltyDraftManager();
 
@@ -408,10 +317,7 @@ public class MiltyDraftHelper {
             String sliceString, List<ComponentSource> allowedSources) {
         try {
             sliceString = sliceString.replace("|", ";");
-            MiltyDraftManager manager = new MiltyDraftManager();
-            manager.init(allowedSources);
-            manager.loadSlicesFromString(sliceString);
-            return manager.getSlices();
+            return DraftSliceHelper.parseSlicesFromString(sliceString);
         } catch (Exception e) {
             BotLogger.error("invalid slice string", e);
             return null;
